feat: Add clean architecture in server dir
