+++ backend/Cargo.toml
@@ -25,3 +25,5 @@ aes-gcm = "0.10.1"
 base64 = "0.21.0"
 axum-extra = "0.7.3"
 tower-http = { version = "0.4.0", features = ["cors"] }
+tracing = "0.1.34"
+tracing-subscriber = "0.3.11"
