+++ backend/src/db/mod.rs
@@ -1,0 +1,3 @@
+pub mod find_dublicate;
+pub mod get_name_and_path_of_file;
+pub mod insert_to_mongo;
