+++ backend/src/tools/generate_uuid.rs
@@ -1,0 +1,3 @@
+pub async fn generate_uuid_v4() -> String {
+    uuid::Uuid::new_v4().to_string()
+}
